+++ test/libsolidity/semanticTests/asmForLoop/for_loop_continue.sol
@@ -1,0 +1,13 @@
+contract C {
+    function f() public returns (uint k) {
+        assembly {
+            for {let i := 0} lt(i, 10) { i := add(i, 1) }
+            {
+                if eq(mod(i, 2), 0) { continue }
+                k := add(k, 1)
+            }
+        }
+    }
+}
+// ----
+// f() -> 5